requirements.txt
flask==3.1.1 
pypdf==5.6.1 
sentence-transformers==4.1.0
langchain==0.3.26
langchain-community==0.3.26
python-dotenv==1.1.0
# langchain-openai==0.3.24
langchain-google-genai
# langchain-google-vertexai==0.2.5
langchain-pinecone==0.2.8 
-e .